requirements-lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
aws-cdk-asset-awscli-v1==2.2.243
aws-cdk-asset-kubectl-v20==2.1.4
aws-cdk-asset-node-proxy-agent-v6==2.1.0
aws-cdk-lib==2.124.0
black==24.1.1
boto3==1.34.34
botocore==1.34.162
cattrs==24.1.3
cfgv==3.4.0
click==8.2.1
constructs==10.3.0
discord-py==2.3.2
distlib==0.3.9
filelock==3.18.0
flake8==7.0.0
frozenlist==1.7.0
identify==2.6.12
idna==3.10
importlib-resources==6.5.2
iniconfig==2.1.0
jmespath==1.0.1
jsii==1.112.0
mccabe==0.7.0
multidict==6.6.3
mypy==1.8.0
mypy-extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
publication==0.0.3
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.2
s3transfer==0.10.4
six==1.17.0
typeguard==2.13.3
typing-extensions==4.14.0
urllib3==2.5.0
virtualenv==20.31.2
yarl==1.20.1
